refactor: move SemesterCard from SemesterList to separate file
