+++ frontend/src/components/Semester/SemesterList/SemesterCopyForm/SemesterCopyForm.js
@@ -6,8 +6,8 @@ import './SemesterCopyForm.scss';
 import {
     FORM_SEMESTER_COPY_RED_HINT,
     FORM_SEMESTER_COPY_HINT,
-} from '../../../constants/translationLabels/formElements';
-import { TEACHER_SCHEDULE_LABEL } from '../../../constants/translationLabels/common';
+} from '../../../../constants/translationLabels/formElements';
+import { TEACHER_SCHEDULE_LABEL } from '../../../../constants/translationLabels/common';
 import SemesterSelect from './SemesterSelect';
 
 const SemesterCopyForm = (props) => {
