+++ frontend/src/components/Semester/SemesterList/SemesterCopyForm/SemesterCopyForm.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/colors.scss';
+@import '../../../../styles/colors.scss';
 
 p.empty_schedule {
     text-align: center;
