+++ frontend/src/components/Semester/SemesterList/SemesterList.scss
@@ -1,4 +1,4 @@
-@import '/../../App.scss';
+@import '../../../App.scss';
 
 .semester-card {
     min-height: 175px;
