+++ frontend/src/containers/SemesterPage/SemesterCopyForm.js
@@ -1,6 +1,6 @@
 import { reduxForm } from 'redux-form';
 import { SEMESTER_COPY_FORM } from '../../constants/reduxForms';
-import SemesterCopyForm from '../../components/Semester/SemesterCopyForm/SemesterCopyForm';
+import SemesterCopyForm from '../../components/Semester/SemesterList/SemesterCopyForm/SemesterCopyForm';
 
 export default reduxForm({
     form: SEMESTER_COPY_FORM,
